% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OECD.R
\docType{data}
\name{OECD}
\alias{OECD}
\title{OECD Dataset}
\format{
A tibble with 6 columns:
\describe{
  \item{Recipient}{Recipient country of the aid}
  \item{Donor}{Donor country}
  \item{Aid type}{Type of aid (e.g., ODA)}
  \item{Amount type}{Type of amount (e.g., Constant Prices)}
  \item{Year}{Year of the data}
  \item{Value}{Amount of aid in millions}
}
}
\source{
OECD
}
\usage{
OECD
}
\description{
This dataset contains aid data from the OECD, detailing aid between donor and recipient countries.
}
\keyword{datasets}
